Move dpc++ helper functions to utils
